refactor: move queries into `api/lib/queries`
